AI-Services/requirements.txt
# Core dependencies for MindMate ChatBot
# Google Gemini AI API client
google-genai>=0.3.0

# Data validation and serialization
pydantic>=2.0.0

# Configuration file parsing
PyYAML>=6.0

# Firebase Admin SDK
firebase-admin>=6.5.0

# FastAPI web framework and ASGI server
fastapi>=0.111.0
uvicorn>=0.30.0

# Standard library imports (no installation needed):
# - typing
# - datetime
# - logging
# - abc
# - json
# - os
# - pathlib

# Optional: For enhanced development and testing
# pytest>=7.0.0
# pytest-asyncio>=0.21.0